+++ instrumentation.go
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2025 The SonicWeb contributors.
+// SPDX-FileCopyrightText: 2026 The SonicWeb contributors.
 // SPDX-License-Identifier: MPL-2.0
 
 package main
@@ -12,102 +12,504 @@ import (
 	"net/http"
 	"net/http/pprof"
 	"os"
+	"strings"
 
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"go.opentelemetry.io/contrib/bridges/otelslog"
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/exporters/prometheus"
-
-	// "go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
+	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploggrpc"
+	"go.opentelemetry.io/otel/exporters/otlp/otlplog/otlploghttp"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
+	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
+	otelprom "go.opentelemetry.io/otel/exporters/prometheus"
+	"go.opentelemetry.io/otel/exporters/stdout/stdoutlog"
+	"go.opentelemetry.io/otel/exporters/stdout/stdoutmetric"
+	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
+	"go.opentelemetry.io/otel/log/global"
 	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/sdk/log"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
-	sdktrace "go.opentelemetry.io/otel/sdk/trace"
-	semconv "go.opentelemetry.io/otel/semconv/v1.27.0"
+	"go.opentelemetry.io/otel/sdk/trace"
+	semconv "go.opentelemetry.io/otel/semconv/v1.37.0"
 )
 
-// initTracer initializes the tracing functionality. If used, the tracing
-// backend will report tracing data to the provided tracing endpoint.
-func initTracer(endpoint string) (*sdktrace.TracerProvider, error) {
-	// // Create stdout exporter to be able to retrieve
-	// // the collected spans.
+const OTLPProtocolGRPC = "grpc"
+const OTLPProtocolHTTP = "http"
+const OTLPExporterConsole = "console"
+const OTLPExporterNone = "none"
+const OTLPExporterOTLP = "otlp"
+const OTLPExporterPrometheus = "prometheus"
 
-	// exporter, err := stdouttrace.New(stdout.WithPrettyPrint())
+var ErrUnsupportedOTLPProtocol = errors.New("unsupported protocol")
 
-	exporter, err := otlptracehttp.New(context.Background(),
-		otlptracehttp.WithCompression(otlptracehttp.GzipCompression),
-		otlptracehttp.WithEndpoint(endpoint),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("could not get telemetry exporter: %w", err)
+// setupOTelSDK bootstraps the OpenTelemetry pipeline.
+// If it does not return an error, make sure to call shutdown for proper cleanup.
+//
+//nolint:funlen // cannot go around the steps, splitting will confuse more than longer function
+func setupOTelSDK(ctx context.Context) (func(context.Context) error, http.Handler, error) {
+	var shutdownFuncs []func(context.Context) error
+	var err error
+
+	// shutdown calls cleanup functions registered via shutdownFuncs.
+	// The errors from the calls are joined.
+	// Each registered cleanup will be invoked once.
+	shutdown := func(ctx context.Context) error {
+		var err error
+		for _, fn := range shutdownFuncs {
+			err = errors.Join(err, fn(ctx))
+		}
+		shutdownFuncs = nil
+
+		return err
 	}
 
-	// For the demonstration, use sdktrace.AlwaysSample sampler to sample all traces.
-	// In a production application, use sdktrace.ProbabilitySampler with a desired probability.
-	tracerProvider := sdktrace.NewTracerProvider(
-		sdktrace.WithBatcher(exporter),
-		sdktrace.WithResource(resource.NewWithAttributes(
-			semconv.SchemaURL,
-			semconv.ServiceName(ServerName))),
-		sdktrace.WithSampler(sdktrace.AlwaysSample()),
+	// handleErr calls shutdown for cleanup and makes sure that all errors are returned.
+	handleErr := func(inErr error) {
+		err = errors.Join(inErr, shutdown(ctx))
+	}
+
+	// Set some environment variables to reasonable values
+	setupEnvDefaults()
+
+	// Set up a propagator.
+	prop := newPropagator()
+	otel.SetTextMapPropagator(prop)
+
+	// Get the common resources to use
+	res, err := newResource(ctx)
+	if err != nil {
+		handleErr(err)
+		return shutdown, nil, err
+	}
+
+	// Set up a trace provider.
+	tracerProvider, err := newTracerProvider(ctx, res)
+	if err != nil {
+		handleErr(err)
+		return shutdown, nil, err
+	}
+
+	if tracerProvider != nil {
+		shutdownFuncs = append(shutdownFuncs, tracerProvider.Shutdown)
+		otel.SetTracerProvider(tracerProvider)
+	}
+
+	// Set up meter provider.
+	meterProvider, metricHandler, err := newMeterProvider(ctx, res)
+	if err != nil {
+		handleErr(err)
+		return shutdown, nil, err
+	}
+
+	if meterProvider != nil {
+		shutdownFuncs = append(shutdownFuncs, meterProvider.Shutdown)
+		otel.SetMeterProvider(meterProvider)
+	}
+
+	// Set up a logger provider.
+	loggerProvider, err := newLoggerProvider(ctx, res)
+	if err != nil {
+		handleErr(err)
+		return shutdown, nil, err
+	}
+
+	if loggerProvider != nil {
+		shutdownFuncs = append(shutdownFuncs, loggerProvider.Shutdown)
+		global.SetLoggerProvider(loggerProvider)
+
+		slog.SetDefault(
+			slog.New(NewMultiHandler(
+				slog.Default().Handler(),
+				otelslog.NewHandler("otel", otelslog.WithLoggerProvider(loggerProvider)))))
+	}
+
+	return shutdown, metricHandler, err
+}
+
+// setupEnvDefaults sets some OpenTelemetry related configuration environment variables to reasonable values.
+func setupEnvDefaults() {
+	defaults := map[string]string{
+		"OTEL_EXPORTER_OTLP_COMPRESSION": "gzip",
+		"OTEL_EXPORTER_OTLP_PROTOCOL":    OTLPProtocolGRPC,
+	}
+
+	for k, v := range defaults {
+		if _, isSet := os.LookupEnv(k); !isSet {
+			if err := os.Setenv(k, v); err != nil {
+				slog.Error("failed to set default environment variable",
+					slog.String("name", k),
+					slog.String("value", v),
+					slog.String("error", err.Error()))
+			}
+		}
+	}
+}
+
+// newPropagator creates a composite propagator based on the environment variable OTEL_PROPAGATORS. Supported
+// propagators depend on the Go OpenTelemetry implementation and are currently limited to TraceContext and Baggage.
+// If none is given, defaults to TraceContext and Baggage together.
+//
+// Environment variable processing:
+// manual:
+//   - OTEL_PROPAGATORS
+//
+//nolint:ireturn // the result is an interface, no choice here
+func newPropagator() propagation.TextMapPropagator {
+	propagators := make([]propagation.TextMapPropagator, 0, 2)
+
+	if envPropagators := os.Getenv("OTEL_PROPAGATORS"); envPropagators != "" {
+		for p := range strings.SplitSeq(envPropagators, ",") {
+			switch strings.TrimSpace(p) {
+			case "baggage":
+				propagators = append(propagators, propagation.Baggage{})
+			case "tracecontext":
+				propagators = append(propagators, propagation.TraceContext{})
+			default:
+				slog.Warn("unsupported propagator in OTEL_PROPAGATORS", slog.String("name", p))
+			}
+		}
+	}
+
+	if len(propagators) == 0 {
+		propagators = append(propagators,
+			propagation.TraceContext{},
+			propagation.Baggage{})
+	}
+
+	return propagation.NewCompositeTextMapPropagator(propagators...)
+}
+
+// newResource configures a resource to be used by the telemetry providers.
+// Attributes configured in the environment variable `OTEL_RESOURCE_ATTRIBUTES` are added.
+// It should contain a comma-separated list of key-value-pairs the form `key0=val0,key1=val1,...`.
+func newResource(ctx context.Context) (*resource.Resource, error) {
+	res, err := resource.New(
+		ctx,
+		resource.WithSchemaURL(semconv.SchemaURL),
+		resource.WithAttributes(
+			semconv.ServiceNameKey.String(ServerName),
+			semconv.ServiceVersionKey.String(buildInfoTag),
+		),
+		resource.WithFromEnv(),
+		resource.WithTelemetrySDK(),
+		resource.WithProcess(),
+		resource.WithContainer(),
+		resource.WithOS(),
+		resource.WithHost(),
 	)
 
-	otel.SetTracerProvider(tracerProvider)
-	otel.SetTextMapPropagator(propagation.NewCompositeTextMapPropagator(
-		propagation.TraceContext{},
-		propagation.Baggage{},
-	))
+	if err != nil {
+		return nil, fmt.Errorf("could not create resource: %w", err)
+	}
+
+	return res, nil
+}
+
+// newTraceExporter initializes a trace.SpanExporter based on the provided exporter name and protocol.
+//
+//nolint:ireturn
+func newTraceExporter(ctx context.Context, name, protocol string) (trace.SpanExporter, error) {
+	var exp trace.SpanExporter
+	var err error
+
+	switch name {
+	case OTLPExporterOTLP:
+		switch protocol {
+		case OTLPProtocolGRPC:
+			exp, err = otlptracegrpc.New(ctx)
+		case OTLPProtocolHTTP:
+			exp, err = otlptracehttp.New(ctx)
+		default:
+			err = ErrUnsupportedOTLPProtocol
+		}
+	case OTLPExporterConsole:
+		exp, err = stdouttrace.New(stdouttrace.WithPrettyPrint())
+	default:
+		slog.Warn("unsupported trace exporter", slog.String("name", name))
+	}
+
+	if err != nil {
+		return nil, fmt.Errorf("error creating trace exporter: %w", err)
+	}
+
+	return exp, nil
+}
+
+// newTracerProvider creates a new trace provider based on the environment variables. For the environment variable
+// `OTEL_TRACES_EXPORTER` it supports the values `otlp`, `console` and `none`, with `none` being the default.
+//
+// Environment variable processing:
+// manual:
+//   - OTEL_EXPORTER_OTLP_PROTOCOL,           OTEL_EXPORTER_OTLP_TRACES_PROTOCOL
+//   - OTEL_TRACES_EXPORTER
+//
+// automatic:
+//   - OTEL_EXPORTER_OTLP_ENDPOINT,           OTEL_EXPORTER_OTLP_TRACES_ENDPOINT
+//   - OTEL_EXPORTER_OTLP_HEADERS,            OTEL_EXPORTER_OTLP_TRACES_HEADERS
+//   - OTEL_EXPORTER_OTLP_TIMEOUT,            OTEL_EXPORTER_OTLP_TRACES_TIMEOUT
+//   - OTEL_EXPORTER_OTLP_COMPRESSION,        OTEL_EXPORTER_OTLP_TRACES_COMPRESSION
+//   - OTEL_EXPORTER_OTLP_CERTIFICATE,        OTEL_EXPORTER_OTLP_TRACES_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE, OTEL_EXPORTER_OTLP_TRACES_CLIENT_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_KEY,         OTEL_EXPORTER_OTLP_TRACES_CLIENT_KEY
+//   - OTEL_TRACES_SAMPLER
+//   - OTEL_TRACES_SAMPLER_ARG
+func newTracerProvider(ctx context.Context, res *resource.Resource) (*trace.TracerProvider, error) {
+	traceExporters := make([]trace.SpanExporter, 0, 1)
+
+	envExporters := os.Getenv("OTEL_TRACES_EXPORTER")
+
+	if envExporters == OTLPExporterNone || envExporters == "" {
+		return nil, nil //nolint:nilnil // it is completely valid to have no provider set
+	}
+
+	protocol := os.Getenv("OTEL_EXPORTER_OTLP_PROTOCOL")
+
+	if specializedProtocol := os.Getenv("OTEL_EXPORTER_OTLP_TRACES_PROTOCOL"); specializedProtocol != "" {
+		protocol = specializedProtocol
+	}
+
+	for exporter := range strings.SplitSeq(envExporters, ",") {
+		exp, err := newTraceExporter(ctx, strings.TrimSpace(exporter), protocol)
+
+		if err != nil {
+			return nil, fmt.Errorf("could not instantiate trace exporter %v with protocol %v: %w",
+				exporter, protocol, err)
+		}
+
+		if exp != nil {
+			traceExporters = append(traceExporters, exp)
+		}
+	}
+
+	tracerProviderOptions := make([]trace.TracerProviderOption, 0, len(traceExporters)+1)
+
+	for _, t := range traceExporters {
+		tracerProviderOptions = append(tracerProviderOptions, trace.WithBatcher(t))
+	}
+
+	if res != nil {
+		tracerProviderOptions = append(tracerProviderOptions, trace.WithResource(res))
+	}
+
+	tracerProvider := trace.NewTracerProvider(tracerProviderOptions...)
 
 	return tracerProvider, nil
 }
 
-// setupMetricsInstrumentation sets up the telemetry functionality.
-// If either telemetry or pprof are enabled, serveMetrics is called subsequently to
-// open a telemetry and/or profiling providing server port.
-func setupMetricsInstrumentation(
-	instrumentAddress *string,
-	instrumentPort *string,
-	enableTelemetry bool,
-	enablePprof bool) {
+// newMeterReader initializes and returns a metric.Exporter, metric.Reader,
+// and optional http.Handler based on the given name and protocol.
+//
+//nolint:ireturn
+func newMeterReader(ctx context.Context, name, protocol string) (metric.Exporter, metric.Reader, http.Handler, error) {
+	var reader metric.Reader
+	var exp metric.Exporter
+	var metricHandler http.Handler
+	var err error
 
-	if !enableTelemetry && !enablePprof {
-		return
+	switch name {
+	case OTLPExporterOTLP:
+		switch protocol {
+		case OTLPProtocolGRPC:
+			exp, err = otlpmetricgrpc.New(ctx)
+		case OTLPProtocolHTTP:
+			exp, err = otlpmetrichttp.New(ctx)
+		default:
+			err = ErrUnsupportedOTLPProtocol
+		}
+	case OTLPExporterPrometheus:
+		reg := prometheus.NewRegistry()                                  // create explicit registry
+		reader, err = otelprom.New(otelprom.WithRegisterer(reg))         // create reader that writes to that registry
+		metricHandler = promhttp.HandlerFor(reg, promhttp.HandlerOpts{}) // create handlers reading exactly that registry
+	case OTLPExporterConsole:
+		exp, err = stdoutmetric.New()
+	default:
+		slog.Warn("unsupported metric exporter", slog.String("name", name))
 	}
 
-	if enableTelemetry {
-		metricsExporter, metricsExporterErr := prometheus.New()
-
-		if metricsExporterErr != nil {
-			slog.Error("could not create prometheus exporter", slog.String("error", metricsExporterErr.Error()))
-			os.Exit(1)
-		}
-
-		baseResource, baseResourceErr := resource.Merge(resource.Default(),
-			resource.NewSchemaless(
-				attribute.KeyValue{
-					Key:   "service.name",
-					Value: attribute.StringValue("sonic-web"),
-				}))
-
-		if baseResourceErr != nil {
-			slog.Error("could not create base resource", slog.String("error", baseResourceErr.Error()))
-		}
-
-		meterProvider := metric.NewMeterProvider(
-			metric.WithReader(metricsExporter),
-			metric.WithResource(baseResource))
-
-		otel.SetMeterProvider(meterProvider)
+	if err != nil {
+		return nil, nil, nil, fmt.Errorf("error creating meter reader: %w", err)
 	}
 
-	go serveMetrics(*instrumentAddress, *instrumentPort, enableTelemetry, enablePprof)
+	return exp, reader, metricHandler, nil
+}
+
+// newMeterProvider creates a new meter provider based on the environment variables.
+// For the environment variable `OTEL_METRICS_EXPORTER` it supports the values `otlp`,
+// `prometheus`, `console` and `none`, with `none` being the default.
+//
+// Environment variable processing:
+// manual:
+//   - OTEL_EXPORTER_OTLP_PROTOCOL,           OTEL_EXPORTER_OTLP_METRICS_PROTOCOL
+//   - OTEL_METRICS_EXPORTER
+//
+// automatic:
+//   - OTEL_EXPORTER_OTLP_ENDPOINT,           OTEL_EXPORTER_OTLP_METRICS_ENDPOINT
+//   - OTEL_EXPORTER_OTLP_HEADERS,            OTEL_EXPORTER_OTLP_METRICS_HEADERS
+//   - OTEL_EXPORTER_OTLP_TIMEOUT,            OTEL_EXPORTER_OTLP_METRICS_TIMEOUT
+//   - OTEL_EXPORTER_OTLP_COMPRESSION,        OTEL_EXPORTER_OTLP_METRICS_COMPRESSION
+//   - OTEL_EXPORTER_OTLP_CERTIFICATE,        OTEL_EXPORTER_OTLP_METRICS_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE, OTEL_EXPORTER_OTLP_METRICS_CLIENT_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_KEY,         OTEL_EXPORTER_OTLP_METRICS_CLIENT_KEY
+//   - OTEL_EXPORTER_OTLP_METRICS_TEMPORALITY_PREFERENCE
+//   - OTEL_EXPORTER_OTLP_METRICS_DEFAULT_HISTOGRAM_AGGREGATION
+//   - OTEL_METRIC_EXPORT_INTERVAL
+//   - OTEL_METRIC_EXPORT_TIMEOUT
+func newMeterProvider(ctx context.Context, res *resource.Resource) (*metric.MeterProvider, http.Handler, error) {
+	metricReaders := make([]metric.Reader, 0, 1)
+
+	envExporters := os.Getenv("OTEL_METRICS_EXPORTER")
+
+	if envExporters == OTLPExporterNone || envExporters == "" {
+		return nil, nil, nil
+	}
+
+	protocol := os.Getenv("OTEL_EXPORTER_OTLP_PROTOCOL")
+
+	if specializedProtocol := os.Getenv("OTEL_EXPORTER_OTLP_METRICS_PROTOCOL"); specializedProtocol != "" {
+		protocol = specializedProtocol
+	}
+
+	var metricHandler http.Handler
+
+	for exporter := range strings.SplitSeq(envExporters, ",") {
+		exp, reader, tmpHandler, err := newMeterReader(ctx, strings.TrimSpace(exporter), protocol)
+
+		if tmpHandler != nil {
+			metricHandler = tmpHandler
+		}
+
+		if err != nil {
+			return nil, nil, fmt.Errorf("could not instantiate metrics exporter %v with protocol %v: %w",
+				exporter, protocol, err)
+		}
+
+		if exp != nil {
+			reader = metric.NewPeriodicReader(exp)
+		}
+
+		if reader != nil {
+			metricReaders = append(metricReaders, reader)
+		}
+	}
+
+	meterProviderOptions := make([]metric.Option, 0, len(metricReaders)+1)
+
+	for _, r := range metricReaders {
+		meterProviderOptions = append(meterProviderOptions, metric.WithReader(r))
+	}
+
+	if res != nil {
+		meterProviderOptions = append(meterProviderOptions, metric.WithResource(res))
+	}
+
+	meterProvider := metric.NewMeterProvider(meterProviderOptions...)
+
+	return meterProvider, metricHandler, nil
+}
+
+// newLoggerExporter creates a log exporter based on the provided name and protocol or returns an error if unsupported.
+//
+//nolint:ireturn
+func newLoggerExporter(ctx context.Context, name, protocol string) (log.Exporter, error) {
+	var exp log.Exporter
+	var err error
+
+	switch name {
+	case OTLPExporterOTLP:
+		switch protocol {
+		case OTLPProtocolGRPC:
+			exp, err = otlploggrpc.New(ctx)
+		case OTLPProtocolHTTP:
+			exp, err = otlploghttp.New(ctx)
+		default:
+			err = ErrUnsupportedOTLPProtocol
+		}
+	case OTLPExporterConsole:
+		exp, err = stdoutlog.New()
+	default:
+		slog.Warn("unsupported log exporter", slog.String("name", name))
+	}
+
+	if err != nil {
+		return nil, fmt.Errorf("error creating logger exporter: %w", err)
+	}
+
+	return exp, nil
+}
+
+// newLoggerProvider creates a new logger provider based on the environment variables. For the environment variable
+// `OTEL_LOGS_EXPORTER` it supports the values `otlp`, `console` and `none`, with `none` being the default.
+//
+// Environment variable processing:
+// manual:
+//   - OTEL_EXPORTER_OTLP_PROTOCOL,           OTEL_EXPORTER_OTLP_LOGS_PROTOCOL
+//   - OTEL_LOGS_EXPORTER
+//
+// automatic:
+//   - OTEL_EXPORTER_OTLP_ENDPOINT,           OTEL_EXPORTER_OTLP_LOGS_ENDPOINT
+//   - OTEL_EXPORTER_OTLP_HEADERS,            OTEL_EXPORTER_OTLP_LOGS_HEADERS
+//   - OTEL_EXPORTER_OTLP_TIMEOUT,            OTEL_EXPORTER_OTLP_LOGS_TIMEOUT
+//   - OTEL_EXPORTER_OTLP_COMPRESSION,        OTEL_EXPORTER_OTLP_LOGS_COMPRESSION
+//   - OTEL_EXPORTER_OTLP_CERTIFICATE,        OTEL_EXPORTER_OTLP_LOGS_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_CERTIFICATE, OTEL_EXPORTER_OTLP_LOGS_CLIENT_CERTIFICATE
+//   - OTEL_EXPORTER_OTLP_CLIENT_KEY,         OTEL_EXPORTER_OTLP_LOGS_CLIENT_KEY
+//   - OTEL_BLRP_SCHEDULE_DELAY
+//   - OTEL_BLRP_MAX_EXPORT_BATCH_SIZE
+//   - OTEL_BLRP_EXPORT_TIMEOUT
+//   - OTEL_BLRP_MAX_QUEUE_SIZE
+//   - OTEL_LOGRECORD_ATTRIBUTE_COUNT_LIMIT
+//   - OTEL_LOGRECORD_ATTRIBUTE_VALUE_LENGTH_LIMIT
+func newLoggerProvider(ctx context.Context, res *resource.Resource) (*log.LoggerProvider, error) {
+	logExporters := make([]log.Exporter, 0, 1)
+
+	envExporters := os.Getenv("OTEL_LOGS_EXPORTER")
+
+	if envExporters == OTLPExporterNone || envExporters == "" {
+		return nil, nil //nolint:nilnil // it is completely valid to have no provider set
+	}
+
+	protocol := os.Getenv("OTEL_EXPORTER_OTLP_PROTOCOL")
+
+	if specializedProtocol := os.Getenv("OTEL_EXPORTER_OTLP_LOGS_PROTOCOL"); specializedProtocol != "" {
+		protocol = specializedProtocol
+	}
+
+	for exporter := range strings.SplitSeq(envExporters, ",") {
+		exp, err := newLoggerExporter(ctx, strings.TrimSpace(exporter), protocol)
+
+		if err != nil {
+			return nil, fmt.Errorf("could not instantiate log exporter %v with protocol %v: %w",
+				exporter, protocol, err)
+		}
+
+		if exp != nil {
+			logExporters = append(logExporters, exp)
+		}
+	}
+
+	loggerProviderOptions := make([]log.LoggerProviderOption, 0, len(logExporters)+1)
+
+	for _, l := range logExporters {
+		loggerProviderOptions = append(loggerProviderOptions, log.WithProcessor(log.NewBatchProcessor(l)))
+	}
+
+	if res != nil {
+		loggerProviderOptions = append(loggerProviderOptions, log.WithResource(res))
+	}
+
+	loggerProvider := log.NewLoggerProvider(loggerProviderOptions...)
+
+	return loggerProvider, nil
 }
 
 // serveMetrics sets up the metrics functionality. It opens a separate port, and metrics collectors can fetch their
 // data from there. For profiling purposes, if enabled, also the pprof endpoints are added on the same port.
-func serveMetrics(address string, port string, enableTelemetry, enablePprof bool) {
-	if !enableTelemetry && !enablePprof {
+func serveMetrics(address string, port string, metricHandler http.Handler, enablePprof bool) {
+	if metricHandler == nil && !enablePprof {
 		return
 	}
 
@@ -126,11 +528,11 @@ func serveMetrics(address string, port string, enableTelemetry, enablePprof bool
 		slog.Info("serving pprof disabled")
 	}
 
-	if enableTelemetry {
-		slog.Info("serving telemetry", slog.String("address", listenAddress+"/metrics"))
-		mux.Handle("GET /metrics", promhttp.Handler())
+	if metricHandler != nil {
+		slog.Info("serving metrics", slog.String("address", listenAddress+"/metrics"))
+		mux.Handle("GET /metrics", metricHandler)
 	} else {
-		slog.Info("serving telemetry disabled")
+		slog.Info("serving metrics disabled")
 	}
 
 	server := http.Server{
@@ -145,7 +547,7 @@ func serveMetrics(address string, port string, enableTelemetry, enablePprof bool
 	go func() {
 		if err := server.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
 			slog.Error("error serving metrics", slog.String("error", err.Error()))
-			os.Exit(1)
+			exitFunc(1)
 		}
 		slog.Info("metrics server stopped to accept new connections")
 	}()